% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_request_functions.R
\name{analytics_request}
\alias{analytics_request}
\title{Analytics Request Function}
\usage{
analytics_request(
  dimensions = NULL,
  metrics = NULL,
  sort = NULL,
  maxResults = NULL,
  filters = NULL,
  startDate = "2000-01-01",
  endDate = Sys.Date()
)
}
\value{
data.frame
}
\description{
This is a wrapper function for the YouTube Analytics API
}
\details{
see https://developers.google.com/youtube/analytics/data_model for API Documentation.
}
\examples{
analytics_request(dimensions = "video", metrics = "views", maxResults = 5)

}
