% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countries.R
\name{channel_countries}
\alias{channel_countries}
\title{Channel Country Data}
\usage{
channel_countries(...)
}
\arguments{
\item{...}{Addt. arguments passed to \code{analytics_request}}
}
\value{
data.frame
}
\description{
Returns top countries by views
}
\examples{
\dontrun{
channel_countries()
}
}
