% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/channel_daily_views.R
\name{channel_daily_views}
\alias{channel_daily_views}
\title{Daily Channel Views}
\usage{
channel_daily_views(...)
}
\arguments{
\item{...}{Additional arguments passed to \code{analytics_request()}}
}
\value{
data.frame
}
\description{
This is a wrapper function for \code{analytics_request}. These are the following
defaults:
\itemize{
\item dimensions: \code{day}
\item metrics: \code{views}
\item startDate: \code{2000-01-01}
\item endDate: \code{Sys.Date()}
}
}
\examples{
channel_daily_views()
channel_daily_views(startDate = "2020-01-01", endDate = "2020-01-31")

}
