% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demographics.R
\name{channel_demographics}
\alias{channel_demographics}
\title{Channel Demographic Data}
\usage{
channel_demographics(...)
}
\arguments{
\item{...}{Addt. arguments passed to \code{analytics_request}}
}
\value{
data.frame
}
\description{
Returns age and gender demographics
}
\examples{
\dontrun{
channel_demographics()
}
}
