% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/channel_details.R
\name{channel_details}
\alias{channel_details}
\title{Channel Details}
\usage{
channel_details(part = "contentDetails", ...)
}
\arguments{
\item{...}{Additional arguments passed to \code{data_channel_request()}.}
}
\value{
data.frame
}
\description{
This is a wrapper function for \code{data_channel_request}. By default it returns
\code{contentDetails} - such as likes, favorites, and uploads playlists.
}
\details{
see https://developers.google.com/youtube/v3/docs/channels/list for more details.
}
\examples{
channel_details()
channel_details(part = "snippet")

}
