% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devices.R
\name{channel_devices}
\alias{channel_devices}
\title{Channel Device Data}
\usage{
channel_devices(...)
}
\arguments{
\item{...}{Addt. arguments passed to \code{analytics_request}}
}
\value{
data.frame
}
\description{
Returns top devices by views
}
\examples{
\dontrun{
channel_devices()
}
}
