% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/channel_live_onDemand.R
\name{channel_live_onDemand}
\alias{channel_live_onDemand}
\title{Channel Live vs On Demand Stats}
\usage{
channel_live_onDemand(...)
}
\arguments{
\item{...}{Additional arguments passed to \code{analytics_request()}.}
}
\value{
data.frame
}
\description{
This is a wrapper function for \code{analytics_request}. These are the following
defaults:
\itemize{
\item dimensions: \code{liveOrOnDemand}
\item metrics: \verb{views,estimatedMinutesWatched,averageViewDuration}
\item startDate: \code{2000-01-01}
\item endDate: \code{Sys.Date()}
}
}
\details{
see https://developers.google.com/youtube/analytics/data_model for more details on dimensions and metrics.
}
\examples{
channel_live_onDemand()
channel_live_onDemand(startDate = "2020-01-01", endDate = "2020-12-31")

}
