% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/channel_playlists.R
\name{channel_playlists}
\alias{channel_playlists}
\title{Channel Playlist}
\usage{
channel_playlists(...)
}
\arguments{
\item{...}{Additional arguments passed to \code{data_playlist_request()}.}
}
\value{
data.frame
}
\description{
This is a wrapper function for \code{data_playlist_request} to return all channel playlists.
By default, the function returns etag, playlistId and itemCount
}
\details{
see https://developers.google.com/youtube/v3/docs/playlists/list for more details.
}
\examples{
channel_playback_location()
channel_playback_location(part = "contentDetails")

}
