% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sharing_services.R
\name{channel_sharing_services}
\alias{channel_sharing_services}
\title{Channel Sharing Services Data}
\usage{
channel_sharing_services(...)
}
\arguments{
\item{...}{Addt. arguments passed to \code{analytics_request}}
}
\value{
data.frame
}
\description{
Returns top sharing services by views
}
\examples{
\dontrun{
channel_sharing_services()
}
}
