% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/channel_stats.R
\name{channel_stats}
\alias{channel_stats}
\title{Channel Stats}
\usage{
channel_stats(...)
}
\arguments{
\item{...}{Additional arguments passed to \code{analytics_request()}.}
}
\value{
data.frame
}
\description{
This is a wrapper function for \code{analytics_request} to return channel statistics.
These are the following defaults:
\itemize{
\item metrics: \verb{views,averageViewDuration,estimatedMinutesWatched,shares,likes,dislikes,comments,subscribersGained,subscribersLost}
\item startDate: \code{2000-01-01}
\item endDate: \code{Sys.Date()}
}
}
\details{
see https://developers.google.com/youtube/analytics/data_model for more details on dimensions and metrics.
}
\examples{
channel_stats()
channel_stats(startDate = "2020-01-01", endDate = "2020-12-31")

}
