% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/channel_subscription_status.R
\name{channel_subscription_status}
\alias{channel_subscription_status}
\title{Channel Subscription Status}
\usage{
channel_subscription_status(...)
}
\arguments{
\item{...}{Additional arguments passed to \code{analytics_request()}.}
}
\value{
data.frame
}
\description{
This is a wrapper function for \code{analytics_request} to return subscription status of channel views.
These are the following defaults:
\itemize{
\item metrics: \verb{views,estimatedMinutesWatched}
\item startDate: \code{2000-01-01}
\item endDate: \code{Sys.Date()}
}
}
\details{
see https://developers.google.com/youtube/analytics/data_model for more details on dimensions and metrics.
}
\examples{
channel_subscription_status()
channel_subscription_status(startDate = "2020-01-01", endDate = "2020-12-31")

}
