% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/channel_top_demographics.R
\name{channel_top_demographics}
\alias{channel_top_demographics}
\title{Channel Top Demographics}
\usage{
channel_top_demographics(...)
}
\arguments{
\item{...}{Additional arguments passed to \code{analytics_request()}.}
}
\value{
data.frame
}
\description{
This is a wrapper function for \code{top_function} to return channel views by age group and gender.
These are the following defaults:
\itemize{
\item dimension: \verb{ageGroup,gender}
\item metrics: \code{viewerPercentage}
\item startDate: \code{2000-01-01}
\item endDate: \code{Sys.Date()}
}
}
\details{
see https://developers.google.com/youtube/analytics/data_model for more details on dimensions and metrics.
}
\examples{
channel_top_demographics()
channel_top_demographics(startDate = "2020-01-01", endDate = "2020-12-31")

}
