% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/channel_top_devices.R
\name{channel_top_devices}
\alias{channel_top_devices}
\title{Channel Top Devices}
\usage{
channel_top_devices(...)
}
\arguments{
\item{...}{Additional arguments passed to \code{analytics_request()}.}
}
\value{
data.frame
}
\description{
This is a wrapper function for \code{top_function} to return channel views by device type.
These are the following defaults:
\itemize{
\item dimension: \code{deviceType}
\item metrics: \verb{views,estimatedMinutesWatched}
\item startDate: \code{2000-01-01}
\item endDate: \code{Sys.Date()}
}
}
\details{
see https://developers.google.com/youtube/analytics/data_model for more details on dimensions and metrics.
}
\examples{
channel_top_devices()
channel_top_devices(startDate = "2020-01-01", endDate = "2020-12-31")

}
