% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/channel_top_external_sources.R
\name{channel_top_external_sources}
\alias{channel_top_external_sources}
\title{Channel Top External Traffic Sources}
\usage{
channel_top_external_sources(...)
}
\arguments{
\item{...}{Additional arguments passed to \code{analytics_request()}.}
}
\value{
data.frame
}
\description{
This is a wrapper function for \code{top_function} to return channel views by external traffic sources.
These are the following defaults:
\itemize{
\item dimension: \code{insightTrafficSourceDetails}
\item metrics: \verb{views,estimatedMinutesWatched}
\item filters: \code{insightTrafficSourceType==EXT_URL}
\item startDate: \code{2000-01-01}
\item endDate: \code{Sys.Date()}
}
}
\details{
see https://developers.google.com/youtube/analytics/data_model for more details on dimensions and metrics.
}
\examples{
channel_top_external_sources()
channel_top_external_sources(startDate = "2020-01-01", endDate = "2020-12-31")

}
