% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/channel_top_sharing_services.R
\name{channel_top_sharing_services}
\alias{channel_top_sharing_services}
\title{Channel Top Sharing Services}
\usage{
channel_top_sharing_services(...)
}
\arguments{
\item{...}{Additional arguments passed to \code{analytics_request()}.}
}
\value{
data.frame
}
\description{
This is a wrapper function for \code{top_function} to return channel shares by sharing services.
These are the following defaults:
\itemize{
\item dimension: \code{sharingService}
\item metrics: \code{shares}
\item startDate: \code{2000-01-01}
\item endDate: \code{Sys.Date()}
}
}
\details{
see https://developers.google.com/youtube/analytics/data_model for more details on dimensions and metrics.
}
\examples{
channel_top_sharing_services()
channel_top_sharing_services(startDate = "2020-01-01", endDate = "2020-12-31")

}
