% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/channel_top_videos.R
\name{channel_top_videos}
\alias{channel_top_videos}
\title{Channel Top Videos}
\usage{
channel_top_videos(...)
}
\arguments{
\item{...}{Additional arguments passed to \code{analytics_request()}.}
}
\value{
data.frame
}
\description{
This is a wrapper function for \code{top_function} to return top viewed channel videos.
These are the following defaults:
\itemize{
\item dimension: \code{video}
\item metrics: \verb{views,estimatedMinutesWatched,shares,likes,dislikes,comments}
\item startDate: \code{2000-01-01}
\item endDate: \code{Sys.Date()}
}
}
\details{
see https://developers.google.com/youtube/analytics/data_model for more details on dimensions and metrics.
}
\examples{
channel_top_videos()
channel_top_videos(startDate = "2020-01-01", endDate = "2020-12-31")

}
