% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traffic_details.R
\name{channel_traffic_details}
\alias{channel_traffic_details}
\title{Channel Traffic Sources Data}
\usage{
channel_traffic_details(trafficType = "EXT_URL", ...)
}
\arguments{
\item{trafficType}{Required. String. Type of traffic source.}

\item{...}{Addt. arguments passed to \code{analytics_request}}
}
\value{
data.frame
}
\description{
Returns top traffic sources by views.  Check out
\url{https://developers.google.com/youtube/analytics/dimensions#Traffic_Source_Dimensions}
for more information.
}
\examples{
\dontrun{
channel_traffic_sources()
}
}
