% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traffic_sources.R
\name{channel_traffic_sources}
\alias{channel_traffic_sources}
\title{Channel Traffic Sources Data}
\usage{
channel_traffic_sources(...)
}
\arguments{
\item{...}{Addt. arguments passed to \code{analytics_request}}
}
\value{
data.frame
}
\description{
Returns top traffic sources by views
}
\examples{
\dontrun{
channel_traffic_sources()
}
}
