% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/channel_videos.R
\name{channel_videos}
\alias{channel_videos}
\title{Channel Top Videos}
\usage{
channel_videos()
}
\value{
data.frame
}
\description{
This is a wrapper function for \code{channel_details} and \code{data_playlistItem_request} to return all channel videos
}
\details{
see https://developers.google.com/youtube/v3/docs/playlistItems/list for more details.
}
\examples{
channel_videos()

}
