% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_request_functions.R
\name{data_channel_request}
\alias{data_channel_request}
\title{Channel Requests}
\usage{
data_channel_request(
  part = NULL,
  categoryId = NULL,
  forUsername = NULL,
  hl = NULL,
  id = NULL,
  managedByMe = NULL,
  maxResults = 5,
  mine = NULL,
  mySubscribers = NULL,
  onBehalfOfContentOwner = NULL,
  pageToken = NULL
)
}
\value{
data.frame
}
\description{
This is a wrapper function for channel data in the YouTube Data API.
}
\details{
see https://developers.google.com/youtube/v3/docs/channels/list for API Documentation.
}
\examples{
data_channel_request(part = "contentDetails", mine = "true")

}
