% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_request_functions.R
\name{data_playlist_request}
\alias{data_playlist_request}
\title{Playlist Requests}
\usage{
data_playlist_request(
  part = NULL,
  channelId = NULL,
  hl = NULL,
  id = NULL,
  maxResults = 5,
  mine = NULL,
  onBehalfOfContentOwner = NULL,
  onBehalfOfContentOwnerChannel = NULL,
  pageToken = NULL
)
}
\value{
data.frame
}
\description{
This is a wrapper function for playlist data in the YouTube Data API.
}
\details{
see https://developers.google.com/youtube/v3/docs/playlists/list for API Documentation.
}
\examples{
data_playlist_request(part = "contentDetails", mine = "true")

}
