% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_request_functions.R
\name{data_video_request}
\alias{data_video_request}
\title{Video Requests}
\usage{
data_video_request(
  part = NULL,
  chart = NULL,
  hl = NULL,
  id = NULL,
  locale = NULL,
  maxHeight = NULL,
  maxResults = NULL,
  maxWidth = NULL,
  myRating = NULL,
  onBehalfOfContentOwner = NULL,
  pageToken = NULL,
  regionCode = NULL,
  videoCategoryId = NULL
)
}
\value{
data.frame
}
\description{
This is a wrapper function for video data in the YouTube Data API.
}
\details{
see https://developers.google.com/youtube/v3/docs/videos/list for API Documentation.
}
\examples{
data_video_request(part = "contentDetails", id = "asdfklj234jknfdsfs")

}
