% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{error_checking}
\alias{error_checking}
\title{Error Check}
\usage{
error_checking(ytResults, id)
}
\arguments{
\item{ytResults}{data.frame. Results from API}

\item{id}{Video or playlistId}
}
\value{
data.frame
}
\description{
Make sure that the data returned from the API is not blank
}
