% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cities.R
\name{playlist_cities}
\alias{playlist_cities}
\title{Playlist City Data}
\usage{
playlist_cities(playlistId = NULL, ...)
}
\arguments{
\item{playlistId}{Required. String. Id of YouTube playlist.}

\item{...}{Addt. arguments passed to \code{analytics_request}}
}
\value{
data.frame
}
\description{
Returns top cities by views
}
\examples{
\dontrun{
playlist_cities(playlistId = "PL2MI040U_GXq1L5JUxNOulWCyXn-7QyZK")
} 
}
