% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countries.R
\name{playlist_countries}
\alias{playlist_countries}
\title{Playlist Country Data}
\usage{
playlist_countries(playlistId = NULL, ...)
}
\arguments{
\item{playlistId}{Required. String. Id of YouTube playlist.}

\item{...}{Addt. arguments passed to \code{analytics_request}}
}
\value{
data.frame
}
\description{
Returns top countries by views
}
\examples{
\dontrun{
playlist_countries(playlistId = "PL2MI040U_GXq1L5JUxNOulWCyXn-7QyZK")
} 
}
