% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/playlist_daily_views.R
\name{playlist_daily_views}
\alias{playlist_daily_views}
\title{Playlist Daily Views}
\usage{
playlist_daily_views(playlistId = NULL, ...)
}
\arguments{
\item{playlistId}{ID of YouTube playlist}

\item{...}{Additional arguments passed to \code{analytics_request()}.}
}
\value{
data.frame
}
\description{
This is a wrapper function for \code{analytics_request} to return daily playlist views.
These are the following defaults:
\itemize{
\item dimension: \code{day}
\item metrics: \code{views}
\item startDate: \code{2000-01-01}
\item endDate: \code{Sys.Date()}
}
}
\details{
see https://developers.google.com/youtube/analytics/data_model for API Documentation.
}
\examples{
playlist_daily_views(playlistId = "asdfljASDFLJ-2348uksdfjl")
playlist_daily_views(playlistId = "asdfljASDFLJ-2348uksdfjl", startDate = "2020-01-01", endDate = "2020-12-31")

}
