% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demographics.R
\name{playlist_demographics}
\alias{playlist_demographics}
\title{Playlist Demographic Data}
\usage{
playlist_demographics(playlistId = NULL, ...)
}
\arguments{
\item{playlistId}{Required. String. Id of YouTube playlist.}

\item{...}{Addt. arguments passed to \code{analytics_request}}
}
\value{
data.frame
}
\description{
Returns age and gender demographics
}
\examples{
\dontrun{
playlist_demographics(playlistId = "PL2MI040U_GXq1L5JUxNOulWCyXn-7QyZK")
} 
}
