% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/playlist_details.R
\name{playlist_details}
\alias{playlist_details}
\title{Playlist Details}
\usage{
playlist_details(playlistId = NULL)
}
\arguments{
\item{playlistId}{ID of YouTube playlist}
}
\value{
data.frame
}
\description{
This is a wrapper function for \code{data_playlist_request} to return etag, publishing date, playlist title, description, and thumbnail info.
}
\details{
see https://developers.google.com/youtube/v3/docs/playlists/list for API Documentation.
}
\examples{
playlist_details(playlistId = "asdfljASDFLJ-2348uksdfjl")

}
