% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/playlist_live_onDemand.R
\name{playlist_live_onDemand}
\alias{playlist_live_onDemand}
\title{Playlist Live vs. On Demand Views}
\usage{
playlist_live_onDemand(playlistId = NULL, ...)
}
\arguments{
\item{playlistId}{ID of YouTube playlist}

\item{...}{Additional arguments passed to \code{analytics_request()}.}
}
\value{
data.frame
}
\description{
This is a wrapper function for \code{analytics_request} to return live vs. on demand data of playlist views.
These are the following defaults:
\itemize{
\item dimension: \code{liveOrOnDemand}
\item metrics: \verb{views,estimatedMinutesWatched,averageViewDuration}
\item startDate: \code{2000-01-01}
\item endDate: \code{Sys.Date()}
}
}
\details{
see https://developers.google.com/youtube/analytics/data_model for API Documentation.
}
\examples{
playlist_live_onDemand(playlistId = "asdfljASDFLJ-2348uksdfjl")
playlist_live_onDemand(playlistId = "asdfljASDFLJ-2348uksdfjl", startDate = "2020-01-01", endDate = "2020-12-31")

}
