% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{playlist_metadata}
\alias{playlist_metadata}
\title{Playlist Metadata}
\usage{
playlist_metadata(playlistId = NULL, ...)
}
\arguments{
\item{playlistId}{Required. String. Id of YouTube Video.}

\item{...}{Addt. arguments passed to \code{data_video_request}}
}
\value{
data.frame
}
\description{
Playlist Metadata
}
\examples{
\dontrun{
playlist_metadata(playlistId = "dQw4w9WgXcQ")
}
}
