% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sharing_services.R
\name{playlist_sharing_services}
\alias{playlist_sharing_services}
\title{Playlist Sharing Services Data}
\usage{
playlist_sharing_services(playlistId = NULL, ...)
}
\arguments{
\item{playlistId}{Required. String. Id of YouTube playlist.}

\item{...}{Addt. arguments passed to \code{analytics_request}}
}
\value{
data.frame
}
\description{
Returns top sharing services by views
}
\examples{
\dontrun{
playlist_sharing_services(playlistId = "PL2MI040U_GXq1L5JUxNOulWCyXn-7QyZK")
} 
}
