% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/playlist_stats.R
\name{playlist_stats}
\alias{playlist_stats}
\title{Playlist Statistics}
\usage{
playlist_stats(playlistId = NULL, ...)
}
\arguments{
\item{playlistId}{ID of YouTube playlist}

\item{...}{Additional arguments passed to \code{analytics_request()}.}
}
\value{
data.frame
}
\description{
This is a wrapper function for \code{analytics_request} to return some high level stats.
These are the following defaults:
\itemize{
\item metrics: \verb{views,averageViewDuration,estimatedMinutesWatched}
\item startDate: \code{2000-01-01}
\item endDate: \code{Sys.Date()}
}
}
\details{
see https://developers.google.com/youtube/analytics/data_model for API Documentation.
}
\examples{
playlist_stats(playlistId = "asdfljASDFLJ-2348uksdfjl")
playlist_stats(playlistId = "asdfljASDFLJ-2348uksdfjl", startDate = "2020-01-01", endDate = "2020-12-31")

}
