% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_period.R
\name{playlist_time_period}
\alias{playlist_time_period}
\title{Playlist Time Period Views}
\usage{
playlist_time_period(playlistId = NULL, period = "day", ...)
}
\arguments{
\item{playlistId}{Required. String. Id of YouTube playlist.}

\item{period}{Required. Time period to breakdown data by. Supported values are
\code{day} and \code{month}}

\item{...}{Addt. arguments passed to \code{analytics_request}}
}
\value{
data.frame
}
\description{
Returns playlist views by day or month.
}
\examples{
\dontrun{
playlist_time_period(playlistId = "PL2MI040U_GXq1L5JUxNOulWCyXn-7QyZK")
} 
}
