% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/playlist_top_countries.R
\name{playlist_top_countries}
\alias{playlist_top_countries}
\title{Playlist Top Countries}
\usage{
playlist_top_countries(playlistId = NULL, ...)
}
\arguments{
\item{playlistId}{ID of YouTube playlist}

\item{...}{Additional arguments passed to \code{analytics_request()}.}
}
\value{
data.frame
}
\description{
This is a wrapper function for \code{top_function} to return playlist views by country.
These are the following defaults:
\itemize{
\item dimension: \code{country}
\item metrics: \verb{views,estimatedMinutesWatched,shares,likes,dislikes,comments}
\item startDate: \code{2000-01-01}
\item endDate: \code{Sys.Date()}
}
}
\details{
see https://developers.google.com/youtube/analytics/data_model for more details on dimensions and metrics.
}
\examples{
playlist_top_countries(playlistId = "asdfljASDFLJ-2348uksdfjl")
playlist_top_countries(playlistId = "asdfljASDFLJ-2348uksdfjl", startDate = "2020-01-01", endDate = "2020-12-31")

}
