% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top_videos.R
\name{playlist_top_videos}
\alias{playlist_top_videos}
\title{Playlist Top Videos}
\usage{
playlist_top_videos(playlistId = NULL, ...)
}
\arguments{
\item{playlistId}{Required. String. Id of YouTube playlist.}

\item{...}{Addt. arguments passed to \code{analytics_request}}
}
\value{
data.frame
}
\description{
Returns top videos by views.
}
\examples{
\dontrun{
playlist_top_videos(playlistId = "PL2MI040U_GXq1L5JUxNOulWCyXn-7QyZK")
} 
}
