% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traffic_details.R
\name{playlist_traffic_details}
\alias{playlist_traffic_details}
\title{Playlist Traffic Sources Data}
\usage{
playlist_traffic_details(playlistId = NULL, trafficType = "EXT_URL", ...)
}
\arguments{
\item{playlistId}{Required. String. Id of YouTube playlist.}

\item{trafficType}{Required. String. Type of traffic source.}

\item{...}{Addt. arguments passed to \code{analytics_request}}
}
\value{
data.frame
}
\description{
Returns top traffic sources by views.  Check out
\url{https://developers.google.com/youtube/analytics/dimensions#Traffic_Source_Dimensions}
for more information.
}
\examples{
\dontrun{
playlist_traffic_sources(playlistId = "PL2MI040U_GXq1L5JUxNOulWCyXn-7QyZK")
} 
}
