% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traffic_sources.R
\name{playlist_traffic_sources}
\alias{playlist_traffic_sources}
\title{Playlist Traffic Sources Data}
\usage{
playlist_traffic_sources(playlistId = NULL, ...)
}
\arguments{
\item{playlistId}{Required. String. Id of YouTube playlist.}

\item{...}{Addt. arguments passed to \code{analytics_request}}
}
\value{
data.frame
}
\description{
Returns top traffic sources by views
}
\examples{
\dontrun{
playlist_traffic_sources(playlistId = "PL2MI040U_GXq1L5JUxNOulWCyXn-7QyZK")
} 
}
