% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/playlist_videos.R
\name{playlist_videos}
\alias{playlist_videos}
\title{Playlist Videos}
\usage{
playlist_videos(playlistId = NULL)
}
\arguments{
\item{playlistId}{ID of YouTube playlist}
}
\value{
data.frame
}
\description{
This is a wrapper function for \code{data_playlistItem_request} to return all playlist videos.
}
\details{
see https://developers.google.com/youtube/v3/docs/playlistItems/list for more details.
}
\examples{
playlist_videos(playlistId = "asdfljASDFLJ-2348uksdfjl")

}
