% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top_function.R
\name{top_function}
\alias{top_function}
\title{Top Function}
\usage{
top_function(dimension, ...)
}
\arguments{
\item{playlistId}{ID of YouTube playlist}
}
\value{
data.frame
}
\description{
This is a wrapper function for \code{analytics_request} to simplify returning the top 10 observations of a dimension.
}
\details{
see https://developers.google.com/youtube/analytics/data_model for more details on dimensions and metrics.
}
