% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/video_audience_retention.R
\name{video_audience_retention}
\alias{video_audience_retention}
\title{Video Audience Retention}
\usage{
video_audience_retention(videoId = NULL, audienceType = "ORGANIC", ...)
}
\arguments{
\item{videoId}{ID of YouTube video.}

\item{...}{Additional arguments passed to \code{analytics_request()}.}
}
\value{
data.frame
}
\description{
This is a wrapper function for \code{analytics_request} to return video audience retention percentages.
These are the following defaults:
\itemize{
\item dimension: \code{elapsedVideoTimeRatio}
\item metrics: \code{audienceWatchRatio}
\item filters: \code{audienceType==ORGANIC}
\item startDate: \code{2000-01-01}
\item endDate: \code{Sys.Date()}
}
}
\details{
see https://developers.google.com/youtube/analytics/data_model for more details on dimensions and metrics.
}
\examples{
video_audience_retention(videoId = "ehd_324sjfk")
video_audience_retention(videoId = "ehd_324sjfkl", startDate = "2020-01-01", endDate = "2020-12-31")

}
