% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/video_captions.R
\name{video_captions}
\alias{video_captions}
\title{Video Captions}
\usage{
video_captions(videoId = NULL)
}
\arguments{
\item{videoId}{ID of YouTube video.}
}
\value{
data.frame
}
\description{
This is a wrapper function for \code{youtube_GET} and \code{httr::GET} to get a video's caption id and then the caption.
}
\details{
see https://developers.google.com/youtube/v3/docs/captions/list for more details.
}
\examples{
video_captions(videoId = "ehd_324sjfk")

}
