% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cities.R
\name{video_cities}
\alias{video_cities}
\title{Video City Data}
\usage{
video_cities(videoId = NULL, ...)
}
\arguments{
\item{videoId}{Required. String. Id of YouTube video.}

\item{...}{Addt. arguments passed to \code{analytics_request}}
}
\value{
data.frame
}
\description{
Returns top cities by views
}
\examples{
\dontrun{
video_cities(videoId = "dQw4w9WgXcQ")
} 
}
