% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countries.R
\name{video_countries}
\alias{video_countries}
\title{Video Country Data}
\usage{
video_countries(videoId = NULL, ...)
}
\arguments{
\item{videoId}{Required. String. Id of YouTube video.}

\item{...}{Addt. arguments passed to \code{analytics_request}}
}
\value{
data.frame
}
\description{
Returns top countries by views
}
\examples{
\dontrun{
video_countries(videoId = "dQw4w9WgXcQ")
} 
}
