% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/video_daily_views.R
\name{video_daily_views}
\alias{video_daily_views}
\title{Daily Video Views}
\usage{
video_daily_views(videoId = NULL, ...)
}
\arguments{
\item{videoId}{ID of YouTube video.}

\item{...}{Additional arguments passed to \code{analytics_request()}.}
}
\value{
data.frame
}
\description{
This is a wrapper function for \code{analytics_request} to return daily video views.
These are the following defaults:
\itemize{
\item dimension: \code{day}
\item metrics: \code{views}
\item startDate: \code{2000-01-01}
\item endDate: \code{Sys.Date()}
}
}
\details{
see https://developers.google.com/youtube/analytics/data_model for more details on dimensions and metrics.
}
\examples{
video_daily_views(videoId = "ehd_324sjfk")
video_daily_views(videoId = "ehd_324sjfkl", startDate = "2020-01-01", endDate = "2020-12-31")

}
