% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/video_details.R
\name{video_details}
\alias{video_details}
\title{Video Details}
\usage{
video_details(videoId = NULL, ...)
}
\arguments{
\item{videoId}{ID of YouTube video.}

\item{...}{Additional arguments passed to \code{data_video_request()}.}
}
\value{
data.frame
}
\description{
This is a wrapper function for \code{data_video_request} to return video details.
By default it returns etags, duration, dimension, definition, and caption status.
}
\details{
see https://developers.google.com/youtube/v3/docs/videos/list for more details.
}
\examples{
video_details(videoId = "ehd_324sjfk")

}
