% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devices.R
\name{video_devices}
\alias{video_devices}
\title{Video Device Data}
\usage{
video_devices(videoId = NULL, ...)
}
\arguments{
\item{videoId}{Required. String. Id of YouTube video.}

\item{...}{Addt. arguments passed to \code{analytics_request}}
}
\value{
data.frame
}
\description{
Returns top devices by views
}
\examples{
\dontrun{
video_devices(videoId = "dQw4w9WgXcQ")
} 
}
