% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/video_live_onDemand.R
\name{video_live_onDemand}
\alias{video_live_onDemand}
\title{Video Live vs On Demand Stats}
\usage{
video_live_onDemand(videoId = NULL, ...)
}
\arguments{
\item{videoId}{ID of YouTube video.}

\item{...}{Additional arguments passed to \code{analytics_request()}.}
}
\value{
data.frame
}
\description{
This is a wrapper function for \code{analytics_request} to return video views broken out by live or on demand status.
These are the following defaults:
\itemize{
\item dimension: \code{liveOrOnDemand}
\item metrics: \verb{views,estimatedMinutesWatched,averageViewDuration}
\item startDate: \code{2000-01-01}
\item endDate: \code{Sys.Date()}
}
}
\details{
see https://developers.google.com/youtube/analytics/data_model for more details on dimensions and metrics.
}
\examples{
video_live_onDemand(videoId = "ehd_324sjfk")
video_live_onDemand(videoId = "ehd_324sjfkl", startDate = "2020-01-01", endDate = "2020-12-31")

}
