% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{video_metadata}
\alias{video_metadata}
\title{Video Metadata}
\usage{
video_metadata(videoId = NULL, ...)
}
\arguments{
\item{videoId}{Required. String. Id of YouTube Video.}

\item{...}{Addt. arguments passed to \code{data_video_request}}
}
\value{
data.frame
}
\description{
Video Metadata
}
\examples{
\dontrun{
video_metadata(videoId = "dQw4w9WgXcQ")
}
}
