% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/video_playback_location.R
\name{video_playback_location}
\alias{video_playback_location}
\title{Video Playback Location}
\usage{
video_playback_location(videoId = NULL, ...)
}
\arguments{
\item{videoId}{ID of YouTube video.}

\item{...}{Additional arguments passed to \code{analytics_request()}.}
}
\value{
data.frame
}
\description{
This is a wrapper function for \code{analytics_request} to return video playback location data.
These are the following defaults:
\itemize{
\item dimension: \code{insightPlaybackLocationType}
\item metrics: \verb{views,estimatedMinutesWatched}
\item startDate: \code{2000-01-01}
\item endDate: \code{Sys.Date()}
}
}
\details{
see https://developers.google.com/youtube/analytics/data_model for more details on dimensions and metrics.
}
\examples{
video_playback_location(videoId = "ehd_324sjfk")
video_playback_location(videoId = "ehd_324sjfkl", startDate = "2020-01-01", endDate = "2020-12-31")

}
