% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sharing_services.R
\name{video_sharing_services}
\alias{video_sharing_services}
\title{Video Sharing Services Data}
\usage{
video_sharing_services(videoId = NULL, ...)
}
\arguments{
\item{videoId}{Required. String. Id of YouTube video.}

\item{...}{Addt. arguments passed to \code{analytics_request}}
}
\value{
data.frame
}
\description{
Returns top sharing services by views
}
\examples{
\dontrun{
video_sharing_services(videoId = "dQw4w9WgXcQ")
} 
}
