% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/video_stats.R
\name{video_stats}
\alias{video_stats}
\title{Video Statistics}
\usage{
video_stats(videoId = NULL, ...)
}
\arguments{
\item{videoId}{ID of YouTube video.}

\item{...}{Additional arguments passed to \code{analytics_request()}.}
}
\value{
data.frame
}
\description{
This is a wrapper function for \code{analytics_request} to return some high level video stats.
These are the following defaults:
\itemize{
\item metrics: \verb{views,averageViewDuration,estimatedMinutesWatched,shares,likes,dislikes,comments,subscribersGained,subscribersLost}
\item startDate: \code{2000-01-01}
\item endDate: \code{Sys.Date()}
}
}
\details{
see https://developers.google.com/youtube/analytics/data_model for more details on dimensions and metrics.
}
\examples{
video_stats(videoId = "ehd_324sjfk")
video_stats(videoId = "ehd_324sjfkl", startDate = "2020-01-01", endDate = "2020-12-31")

}
