% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/video_subscription_status.R
\name{video_subscription_status}
\alias{video_subscription_status}
\title{Video Subscription Status}
\usage{
video_subscription_status(videoId = NULL, ...)
}
\arguments{
\item{videoId}{ID of YouTube video.}

\item{...}{Additional arguments passed to \code{analytics_request()}.}
}
\value{
data.frame
}
\description{
This is a wrapper function for \code{analytics_request} to return subscription status of video views.
These are the following defaults:
\itemize{
\item dimensions: \code{subscribedStatus}
\item metrics: \verb{views,estimatedMinutesWatched}
\item startDate: \code{2000-01-01}
\item endDate: \code{Sys.Date()}
}
}
\details{
see https://developers.google.com/youtube/analytics/data_model for more details on dimensions and metrics.
}
\examples{
video_subscription_status(videoId = "ehd_324sjfk")
video_subscription_status(videoId = "ehd_324sjfkl", startDate = "2020-01-01", endDate = "2020-12-31")

}
