% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/video_top_demographics.R
\name{video_top_demographics}
\alias{video_top_demographics}
\title{Video Top Demographics}
\usage{
video_top_demographics(videoId = NULL, ...)
}
\arguments{
\item{videoId}{ID of YouTube video.}

\item{...}{Additional arguments passed to \code{analytics_request()}.}
}
\value{
data.frame
}
\description{
This is a wrapper function for \code{top_function} to return video views by age & gender.
These are the following defaults:
\itemize{
\item dimensions: \verb{ageGroup, gender}
\item metrics: \code{viewerPercentage}
\item startDate: \code{2000-01-01}
\item endDate: \code{Sys.Date()}
}
}
\details{
see https://developers.google.com/youtube/analytics/data_model for more details on dimensions and metrics.
}
\examples{
video_top_demographics(videoId = "ehd_324sjfk")
video_top_demographics(videoId = "ehd_324sjfkl", startDate = "2020-01-01", endDate = "2020-12-31")

}
