% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traffic_details.R
\name{video_traffic_details}
\alias{video_traffic_details}
\title{Video Traffic Details Data}
\usage{
video_traffic_details(videoId = NULL, trafficType = "EXT_URL", ...)
}
\arguments{
\item{videoId}{Required. String. Id of YouTube video.}

\item{trafficType}{Required. String. Type of traffic source.}

\item{...}{Addt. arguments passed to \code{analytics_request}}
}
\value{
data.frame
}
\description{
Returns traffic details by views. Check out
\url{https://developers.google.com/youtube/analytics/dimensions#Traffic_Source_Dimensions}
for more information.
}
\examples{
\dontrun{
video_traffic_sources(videoId = "dQw4w9WgXcQ")
} 
}
