% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traffic_sources.R
\name{video_traffic_sources}
\alias{video_traffic_sources}
\title{Video Traffic Sources Data}
\usage{
video_traffic_sources(videoId = NULL, ...)
}
\arguments{
\item{videoId}{Required. String. Id of YouTube video.}

\item{...}{Addt. arguments passed to \code{analytics_request}}
}
\value{
data.frame
}
\description{
Returns top traffic sources by views
}
\examples{
\dontrun{
video_traffic_sources(videoId = "dQw4w9WgXcQ")
} 
}
