% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_request_functions.R
\name{youtube_GET}
\alias{youtube_GET}
\title{YouTube Get Function}
\usage{
youtube_GET(url, request, token = getOption(file))
}
\value{
data.frame
}
\description{
This is a helper function to execute GET requests and clean the data depending on what API is being called
}
\examples{
youtube_GET("https://www.google.com", request = "data")

}
