% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtube_oauth.R
\name{youtube_oauth}
\alias{youtube_oauth}
\title{YouTube API OAuth}
\usage{
youtube_oauth(appId = NULL, appSecret = NULL, token = NULL)
}
\arguments{
\item{appId}{character string}

\item{appSecret}{character string}

\item{token}{environment}
}
\value{
token environment
}
\description{
This creates or grabs token to authorize API requests
}
\examples{
youtube_oauth("asdoiasdufhoakljfh", "hafodsifuebwmnbhk", ".httr-oauth")

}
