% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{analytics_request}
\alias{analytics_request}
\title{Analytics Request Function}
\usage{
analytics_request(
  dimensions = NULL,
  metrics = NULL,
  sort = NULL,
  maxResults = NULL,
  filters = NULL,
  startDate = "2000-01-01",
  endDate = Sys.Date(),
  ids = "channel==MINE",
  currency = NULL,
  startIndex = NULL,
  includeHistoricalChannelData = NULL,
  token = getOption("YouTube_Token")
)
}
\arguments{
\item{dimensions}{String. See \url{https://developers.google.com/youtube/analytics/dimensions}
for valid arguments.}

\item{metrics}{}

\item{sort}{}

\item{maxResults}{}

\item{filters}{}

\item{startDate}{}

\item{endDate}{}

\item{ids}{}

\item{currency}{}

\item{startIndex}{}

\item{includeHistoricalChannelData}{}

\item{token}{}
}
\description{
This is the function to make API calls to the YouTube Analytics API. Most other
functions in this package are wrappers for this function with some arguments already populated.
}
