% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cities.R
\name{channel_cities}
\alias{channel_cities}
\title{Channel City Data}
\usage{
channel_cities(...)
}
\arguments{
\item{...}{Addt. arguments passed to \code{analytics_request}}
}
\value{
data.frame
}
\description{
Returns top cities by views
}
\examples{
\dontrun{
channel_cities()
}
}
