% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cities.R, R/demographics.R
\name{channel_cities}
\alias{channel_cities}
\title{Channel City Data}
\usage{
channel_cities(...)

channel_cities(...)
}
\arguments{
\item{...}{Addt. arguments passed to \code{analytics_request}}
}
\value{
data.frame

data.frame
}
\description{
Returns top cities by views

Returns age and gender demographics
}
\examples{
\dontrun{
channel_cities()
}
\dontrun{
channel_demographics()
}
}
