% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/live_onDemand.R
\name{channel_live_onDemand}
\alias{channel_live_onDemand}
\title{Channel Live & OnDemand Data}
\usage{
channel_live_onDemand(...)
}
\arguments{
\item{...}{Addt. arguments passed to \code{analytics_request}}
}
\value{
data.frame()
}
\description{
Return channel data split by live & on demand.
}
\examples{
\dontrun{
channel_live_onDemand()
}
}
