% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/playlist_location.R
\name{channel_playback_location}
\alias{channel_playback_location}
\title{Channel Playback Location}
\usage{
channel_playback_location(...)
}
\arguments{
\item{...}{Addt. arguments passed to \code{analytics_request}}
}
\value{
data.frame
}
\description{
Returns the type of page or application where video playbacks occurred.
For more information, refer to
\url{https://developers.google.com/youtube/analytics/dimensions#Playback_Location_Dimensions}
}
\examples{
\dontrun{
channel_playlist_location()
}
}
