% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/playlists.R, R/videos.R
\name{channel_playlists}
\alias{channel_playlists}
\title{Channel Playlists}
\usage{
channel_playlists(...)

channel_playlists(...)
}
\arguments{
\item{...}{Addt. arguments passed to \code{data_playlist_request}}
}
\value{
data.frame

data.frame
}
\description{
Returns all playlists in channel.
}
\examples{
\dontrun{
channel_playlists()
}
\dontrun{
playlist_videos(playlistId = "PL2MI040U_GXq1L5JUxNOulWCyXn-7QyZK")
} 
}
