% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/playlists.R
\name{channel_playlists}
\alias{channel_playlists}
\title{Channel Playlists}
\usage{
channel_playlists(...)
}
\arguments{
\item{...}{Addt. arguments passed to \code{data_playlist_request}}
}
\value{
data.frame
}
\description{
Returns all playlists in channel.
}
\examples{
\dontrun{
channel_playlists()
}
}
