% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stats.R
\name{channel_stats}
\alias{channel_stats}
\title{Channel Statistics}
\usage{
channel_stats(...)
}
\arguments{
\item{...}{Addt. arguments passed to \code{analytics_request}}
}
\value{
data.frame
}
\description{
#' @description
Returns high level video stats.
}
\examples{
\dontrun{
channel_stats()
}
}
