% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subscription_status.R
\name{channel_subscription_status}
\alias{channel_subscription_status}
\title{Channel Subscription Status}
\usage{
channel_subscription_status(...)
}
\arguments{
\item{...}{Addt. arguments passed to \code{analytics_request}}
}
\value{
data.frame
}
\description{
#' @description
Returns the subscription status of channel views.
}
\examples{
\dontrun{
channel_subscription_status()
}
}
