% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top_videos.R
\name{channel_top_videos}
\alias{channel_top_videos}
\title{Channel Top Videos}
\usage{
channel_top_videos(...)
}
\arguments{
\item{...}{Addt. arguments passed to \code{analytics_request}}
}
\value{
data.frame
}
\description{
Returns top videos by views
}
\examples{
\dontrun{
channel_top_videos()
}
}
