% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/videos.R
\name{channel_videos}
\alias{channel_videos}
\title{Channel Videos List}
\usage{
channel_videos(...)
}
\arguments{
\item{...}{Addt. arguments passed to \code{analytics_request}}
}
\value{
data.frame
}
\description{
Returns all videos in channel. Note: As of July 2023, the YouTube Data API
fails to return ALL videos on a channel by getting all videos from the "uploads" playlist
for large YouTube channels. As a result this function's work-around is to pull
all additional playlists from the channel, pull all video lists from all playlists,
and append the result to the uploads return.
}
\examples{
\dontrun{
channel_videos()
}
}
