% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{data_channel_request}
\alias{data_channel_request}
\title{Title}
\usage{
data_channel_request(
  part = NULL,
  categoryId = NULL,
  forUsername = NULL,
  hl = NULL,
  id = NULL,
  managedByMe = NULL,
  maxResults = 5,
  mine = NULL,
  mySubscribers = NULL,
  onBehalfOfContentOwner = NULL,
  pageToken = NULL,
  token = getOption("YouTube_Token")
)
}
\arguments{
\item{part}{}

\item{categoryId}{}

\item{forUsername}{}

\item{hl}{}

\item{id}{}

\item{managedByMe}{}

\item{maxResults}{}

\item{mine}{}

\item{mySubscribers}{}

\item{onBehalfOfContentOwner}{}

\item{pageToken}{}

\item{token}{}
}
\description{
Title
}
