% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{data_playlistItem_request}
\alias{data_playlistItem_request}
\title{Title}
\usage{
data_playlistItem_request(
  part = NULL,
  id = NULL,
  maxResults = 5,
  onBehalfOfContentOwner = NULL,
  pageToken = NULL,
  playlistId = NULL,
  videoId = NULL,
  token = getOption("YouTube_Token")
)
}
\arguments{
\item{part}{}

\item{id}{}

\item{maxResults}{}

\item{onBehalfOfContentOwner}{}

\item{pageToken}{}

\item{playlistId}{}

\item{videoId}{}

\item{token}{}
}
\description{
Title
}
