% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{data_video_request}
\alias{data_video_request}
\title{Title}
\usage{
data_video_request(
  part = NULL,
  chart = NULL,
  hl = NULL,
  id = NULL,
  locale = NULL,
  maxHeight = NULL,
  maxResults = NULL,
  maxWidth = NULL,
  myRating = NULL,
  onBehalfOfContentOwner = NULL,
  pageToken = NULL,
  regionCode = NULL,
  videoCategoryId = NULL,
  token = getOption("YouTube_Token")
)
}
\arguments{
\item{part}{}

\item{chart}{}

\item{hl}{}

\item{id}{}

\item{locale}{}

\item{maxHeight}{}

\item{maxResults}{}

\item{maxWidth}{}

\item{myRating}{}

\item{onBehalfOfContentOwner}{}

\item{pageToken}{}

\item{regionCode}{}

\item{videoCategoryId}{}

\item{token}{}
}
\description{
Title
}
