% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{error_checking}
\alias{error_checking}
\title{Error Check}
\usage{
error_checking(ytResults, id, contentType)
}
\arguments{
\item{ytResults}{data.frame. Results from API}

\item{id}{string. Video or playlistId}

\item{contentType}{string. Either "video" or "playlist"}
}
\value{
data.frame
}
\description{
Make sure that the data returned from the API is not blank
}
