% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/live_onDemand.R
\name{playlist_live_onDemand}
\alias{playlist_live_onDemand}
\title{Playlist Live & OnDemand Data}
\usage{
playlist_live_onDemand(playlistId = NULL, ...)
}
\arguments{
\item{playlistId}{Required. String. Id of YouTube playlist.}

\item{...}{Addt. arguments passed to \code{analytics_request}}
}
\value{
data.frame
}
\description{
Return playlist data split by live & on demand.
}
\examples{
\dontrun{
playlist_live_onDemand(playlistId = "PL2MI040U_GXq1L5JUxNOulWCyXn-7QyZK")
}
}
