% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/playback_location.R
\name{playlist_playback_location}
\alias{playlist_playback_location}
\title{Playlist Playback Location}
\usage{
playlist_playback_location(playlistId = NULL, ...)
}
\arguments{
\item{playlistId}{Required. String. Id of YouTube playlist.}

\item{...}{Addt. arguments passed to \code{analytics_request}}
}
\value{
data.frame
}
\description{
Returns the type of page or application where video playbacks occurred.
}
\details{
For more information, refer to
\url{https://developers.google.com/youtube/analytics/dimensions#Playback_Location_Dimensions}
}
\examples{
\dontrun{
playlist_playback_location(playlistId = "PL2MI040U_GXq1L5JUxNOulWCyXn-7QyZK")
}
}
