% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stats.R
\name{playlist_stats}
\alias{playlist_stats}
\title{Playlist Statistics}
\usage{
playlist_stats(playlistId = NULL, ...)
}
\arguments{
\item{playlistId}{Required. Strings. Id of YouTube Playlist}

\item{...}{Addt. arguments passed to \code{analytics_request}}
}
\value{
data.frame
}
\description{
#' @description
Returns high level video stats.
}
\examples{
\dontrun{
playlist_stats(playlistId = "PL2MI040U_GXq1L5JUxNOulWCyXn-7QyZK")
}
}
