% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subscription_status.R
\name{playlist_subscription_status}
\alias{playlist_subscription_status}
\title{Title}
\usage{
playlist_subscription_status(playlistId = NULL, ...)
}
\arguments{
\item{playlistId}{Required. String. Id of YouTube Playlist}

\item{...}{Addt. arguments passed to \code{analytics_request}}
}
\value{
data.frame
}
\description{
#' @description
Returns the subscription status of playlist views.
}
\examples{
\dontrun{
playlist_subscription_status(playlistId = "PL2MI040U_GXq1L5JUxNOulWCyXn-7QyZK")
} 
}
