% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/videos.R
\name{playlist_videos}
\alias{playlist_videos}
\title{Playlist Videos List}
\usage{
playlist_videos(playlistId = NULL, ...)
}
\arguments{
\item{playlistId}{Required. String. Id of YouTube playlist.}

\item{...}{Addt. arguments passed to \code{analytics_request}}
}
\value{
data.frame
}
\description{
Returns all videos in playlist
}
\examples{
\dontrun{
playlist_videos(playlistId = "PL2MI040U_GXq1L5JUxNOulWCyXn-7QyZK")
} 
}
