% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audience_retention.R
\name{video_audience_retention}
\alias{video_audience_retention}
\title{Video Audience Retention}
\usage{
video_audience_retention(videoId = NULL, audienceType = "ORGANIC", ...)
}
\arguments{
\item{videoId}{Required. String. Id of YouTube video.}

\item{audienceType}{Required. String. Supported values are \code{ORGANIC},
\code{AD_INSTREAM}, and \code{AD_INDISPLAY}}

\item{...}{Addt. arguments passed to \code{analytics_request}}
}
\value{
data.frame
}
\description{
Returns audience retention percentages for a video. By default this function will
return audience retention rates for organic views.
}
\examples{
\dontrun{
video_audience_retention(videoId = "dQw4w9WgXcQ")
} 
}
