% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/playback_location.R
\name{video_playback_location}
\alias{video_playback_location}
\title{Video Playback Location}
\usage{
video_playback_location(videoId = NULL, ...)
}
\arguments{
\item{videoId}{Required. String. Id of YouTube video.}

\item{...}{Addt. arguments passed to \code{analytics_request}}
}
\value{
data.frame
}
\description{
Returns the type of page or application where video playbacks occurred.
}
\details{
For more information, refer to
\url{https://developers.google.com/youtube/analytics/dimensions#Playback_Location_Dimensions}
}
\examples{
\dontrun{
video_playback_location(videoId = "dQw4w9WgXcQ")
}
}
