% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stats.R
\name{video_stats}
\alias{video_stats}
\title{Video Statistics}
\usage{
video_stats(videoId = NULL, ...)
}
\arguments{
\item{videoId}{Required. String. Id of YouTube video.}

\item{...}{Addt. arguments passed to \code{analytics_request}}
}
\value{
data.frame
}
\description{
Returns high level video stats.
}
\examples{
\dontrun{
video_stats(videoId = "dQw4w9WgXcQ")
} 
}
