% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subscription_status.R
\name{video_subscription_status}
\alias{video_subscription_status}
\title{Video Subscription Status}
\usage{
video_subscription_status(videoId = NULL, ...)
}
\arguments{
\item{videoId}{Required. String. Id of YouTube video.}

\item{...}{Addt. arguments passed to \code{analytics_request}}
}
\value{
data.frame
}
\description{
Returns the subscription status of video views.
}
\examples{
\dontrun{
video_subscription_status(videoId = "dQw4w9WgXcQ")
} 
}
