% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{youtube_GET}
\alias{youtube_GET}
\title{YouTube GET Function}
\usage{
youtube_GET(url = NULL, request = NULL, token)
}
\arguments{
\item{url}{Required. Url path for API request}

\item{request}{Required. Used to specify if requesting data from the
YouTube Analytics API or YouTube Data api. Either 'analytics' or 'data'.}

\item{token}{Required.}
}
\value{
data.frame
}
\description{
this is a helper function to execute GET requests and clean data
depending on what API is being called.
This function can be but is not intended to be called directly.
}
\examples{
\dontrun{
youtube_GET(paste0("https://youtubeanalytics.googleapis.com/v2/reports?",
                   "endDate=2018-05-01",
                   "&ids=channel==MINE",
                   "&metrics=views,comments,likes,dislikes,estimatedMinutesWatched",
                   "&startDate=2017-01-01"),
           "analytics")
}
}
